+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Notable changes of tdmclient-ty. Release versions refer to [https://pypi.org/pro
 
 ## [Unreleased]
 
+- Warnings for local variables which hide global variables (a declaration as global could be missing)
+
 ## [0.1.2] - 2022-02-23
 
 ### Added